pyproject.toml
[tool.poetry]
name = "trimmothy"
version = "0.1.0"
description = "A simple and elegant video trimming application for macOS"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "trimmothy", from = "src"}]

[tool.poetry.dependencies]
python = "^3.13"
customtkinter = "^5.2.0"
moviepy = "^1.0.3"
opencv-python = "^4.8.0"
pillow = "^10.0.0"
numpy = "^1.24.0"

[tool.poetry.group.dev.dependencies]
# Add development dependencies here if needed

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
trimmothy = "trimmothy.main:main"
